feat: Enable Global Menu in AF
