+++ packages/frontend/src/featureFlags/FeatureFlags.ts
@@ -9,6 +9,7 @@ export interface FeatureFlagDefinition {
 export const featureFlagDefinitions = [
   { key: 'party.stopReversingPersonNameOrder', type: 'boolean', default: false },
   { key: 'globalMenu.enableAccessManagementLink', type: 'boolean', default: false },
+  { key: 'globalMenu.enabled', type: 'boolean', default: false },
 ] as const satisfies readonly FeatureFlagDefinition[];
 
 export type FeatureFlagKey = (typeof featureFlagDefinitions)[number]['key'];
